There's a clock now that you can touch to show a message.

+++ LostMinds/soundComponent.h
@@ -47,8 +47,8 @@ public:
 		}
 	}
 	void update() override{
-		playerX = Game::camera.x+576;
-		playerY = Game::camera.y+320;
+		playerX = (float)Game::camera.x+576;
+		playerY = (float)Game::camera.y+320;
 		SoundManager::playEffect(effect, getAngle(), getDistance());
 	}
 	/*
